feat(pack.migration): rework on building details pto input (#3319)
* refactor(pack.migration): refactoring code for xdsl pack migration
* feat(pack.migration): rework on building pto to have less user errors
* feat(pack.migration): update from feedback comments
* feat(xdsl.migration): user static method when is needed
* feat(pack.migration): rework on building pto to have less user errors
* feat(pack.migration): update from feedback comments
* feat(xdsl.migrations): use oui-back-button
* feat(xdsl.migration): add style into index.js
* feat(xdsl.migration): remove pack name routing and resolve typo
* feat(xdsl.migration): use constant and catch error

ref: UXCT-300
