+++ packages/manager/apps/telecom/src/app/app.less
@@ -103,7 +103,6 @@
 @import './common/less/navbar.less';
 @import './common/less/variables.less';
 @import './common/less/mixins.less';
-@import './telecom/pack/migration/telecom-pack-migration.less';
 @import './telecom/pack/move/pack-move.less';
 @import './telecom/pack/resiliation/pack-resiliation.less';
 @import './telecom/pack/slots/voipLine/activation/pack-voipLine-activation.less';
