+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/building-details/migration-building-details.component.js
@@ -1,0 +1,7 @@
+import controller from './migration-building-details.controller';
+import template from './migration-building-details.html';
+
+export default {
+  controller,
+  template,
+};
