+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/building-details/migration-building-details.constant.js
@@ -1,0 +1,11 @@
+export const FIBER_PTO = {
+  FIBER_PTO_YES: 'yes',
+  FIBER_PTO_YES_BUT_NOT_KNOWN: 'yesNotKnown',
+  FIBER_PTO_NO: 'no',
+};
+
+export const STAIR_FLOOR = {
+  unknown: '_NA_',
+};
+
+export default { FIBER_PTO, STAIR_FLOOR };
