+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/confirm/index.js
@@ -1,0 +1,15 @@
+import angular from 'angular';
+
+import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
+import uiRouter from '@uirouter/angularjs';
+import angularTranslate from 'angular-translate';
+
+import component from './pack-migration-confirm.component';
+
+const moduleName = 'ovhManagerTelecomPackMigrationConfirm';
+
+angular
+  .module(moduleName, [ngTranslateAsyncLoader, uiRouter, angularTranslate])
+  .component('packMigrationConfirm', component);
+
+export default moduleName;
