+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/confirm/pack-migration-confirm.component.js
@@ -1,0 +1,7 @@
+import controller from './pack-migration-confirm.controller';
+import template from './pack-migration-confirm.html';
+
+export default {
+  controller,
+  template,
+};
