+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/migration/index.js
@@ -1,0 +1,15 @@
+import angular from 'angular';
+
+import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
+import uiRouter from '@uirouter/angularjs';
+import angularTranslate from 'angular-translate';
+
+import component from './pack-migration-migration.component';
+
+const moduleName = 'ovhManagerTelecomPackMigrationMigration';
+
+angular
+  .module(moduleName, [ngTranslateAsyncLoader, uiRouter, angularTranslate])
+  .component('packMigrationMigration', component);
+
+export default moduleName;
