+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/migration/pack-migration-migration.component.js
@@ -1,0 +1,7 @@
+import controller from './pack-migration-migration.controller';
+import template from './pack-migration-migration.html';
+
+export default {
+  controller,
+  template,
+};
