+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/migration/pack-migration-migration.html
@@ -2,28 +2,28 @@
     <div
         class="alert alert-info"
         role="alert"
-        data-ng-if="MigrationCtrl.loading.migrate && MigrationCtrl.process.migrationDoing"
+        data-ng-if="$ctrl.loading.migrate && $ctrl.process.migrationDoing"
         span
         data-translate="telecom_pack_migration_already_pending"
     ></div>
 
-    <div class="text-center" data-ng-if="MigrationCtrl.loading.migrate">
+    <div class="text-center" data-ng-if="$ctrl.loading.migrate">
         <oui-spinner></oui-spinner>
         <h2 data-translate="telecom_pack_migration_doing"></h2>
     </div>
 
-    <div data-ng-if="!MigrationCtrl.loading.migrate">
+    <div data-ng-if="!$ctrl.loading.migrate">
         <div
             class="alert alert-success"
             role="alert"
-            data-ng-if="MigrationCtrl.migrationStatus === 'success'"
+            data-ng-if="$ctrl.migrationStatus === 'success'"
             data-translate="telecom_pack_migration_success_title"
         ></div>
 
         <div
             class="alert alert-danger"
             role="alert"
-            data-ng-if="MigrationCtrl.migrationStatus === 'error'"
+            data-ng-if="$ctrl.migrationStatus === 'error'"
             data-translate="telecom_pack_migration_error"
         ></div>
     </div>
