+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/offers/index.js
@@ -1,0 +1,15 @@
+import angular from 'angular';
+
+import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
+import uiRouter from '@uirouter/angularjs';
+import angularTranslate from 'angular-translate';
+
+import component from './pack-migration-offers.component';
+
+const moduleName = 'ovhManagerTelecomPackMigrationOffers';
+
+angular
+  .module(moduleName, [ngTranslateAsyncLoader, uiRouter, angularTranslate])
+  .component('packMigrationOffers', component);
+
+export default moduleName;
