+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/offers/pack-migration-offers.component.js
@@ -1,0 +1,7 @@
+import controller from './pack-migration-offers.controller';
+import template from './pack-migration-offers.html';
+
+export default {
+  controller,
+  template,
+};
