+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/pack-migration.component.js
@@ -1,0 +1,11 @@
+import controller from './pack-migration.controller';
+import template from './pack-migration.html';
+
+export default {
+  controller,
+  template,
+  bindings: {
+    packName: '<',
+    goBack: '<',
+  },
+};
