+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/pack-migration.less
@@ -1,3 +1,5 @@
+@import '~ovh-ui-kit-bs/dist/less/bootstrap/_variables';
+
 .telecom-pack-migration {
   .table-comparison {
     &__property {
