+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/pack-migration.routing.js
@@ -1,0 +1,11 @@
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.migration', {
+    url: '/migration',
+    views: {
+      'packView@telecom.packs': 'packMigration',
+    },
+    resolve: {
+      goBack: /* @ngInject */ ($state) => () => $state.go('telecom.packs.pack'),
+    },
+  });
+};
