+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/service-delete/pack-migration-service-delete.component.js
@@ -1,0 +1,7 @@
+import controller from './pack-migration-service-delete.controller';
+import template from './pack-migration-service-delete.html';
+
+export default {
+  controller,
+  template,
+};
