+++ packages/manager/apps/telecom/src/app/telecom/pack/migration/shipping/pack-migration-shipping.component.js
@@ -1,0 +1,7 @@
+import controller from './pack-migration-shipping.controller';
+import template from './pack-migration-shipping.html';
+
+export default {
+  controller,
+  template,
+};
