--- packages/manager/apps/telecom/src/app/telecom/pack/migration/telecom-pack-migration.html
@@ -1,54 +1,0 @@
-<div class="telecom-pack-migration">
-    <header>
-        <tuc-section-back-link
-            data-tuc-section-back-link-to-state="telecom.packs.pack"
-            data-tuc-section-back-link-title="{{ 'telecom_pack_migration_back_link_title' | translate }}"
-        ></tuc-section-back-link>
-        <h1 data-translate="telecom_pack_migration_title"></h1>
-    </header>
-
-    <div class="telecom-section-content">
-        <!-- LOADER -->
-        <div class="text-center" data-ng-if="PackMigrationCtrl.loading.init">
-            <oui-spinner></oui-spinner>
-        </div>
-
-        <tuc-toast-message></tuc-toast-message>
-
-        <!-- OFFER CHOICE STEP -->
-        <div
-            data-ng-if="!PackMigrationCtrl.loading.init && PackMigrationCtrl.process.currentStep === 'offers'"
-            data-ui-view="offersView"
-        ></div>
-
-        <!-- BUILDING DETAILS STEP -->
-        <div
-            data-ng-if="!PackMigrationCtrl.loading.init && PackMigrationCtrl.process.currentStep === 'buildingDetails'"
-            data-ui-view="buildingDetailsView"
-        ></div>
-
-        <!-- SHIPPING STEP -->
-        <div
-            data-ng-if="!PackMigrationCtrl.loading.init && PackMigrationCtrl.process.currentStep === 'shipping'"
-            data-ui-view="shippingView"
-        ></div>
-
-        <!-- CONFIRM STEP -->
-        <div
-            data-ng-if="!PackMigrationCtrl.loading.init && PackMigrationCtrl.process.currentStep === 'confirm'"
-            data-ui-view="confirmView"
-        ></div>
-
-        <!-- MIGRATION TASK IN PROGRESS -->
-        <div
-            data-ng-if="!PackMigrationCtrl.loading.init && PackMigrationCtrl.process.currentStep === 'migration'"
-            data-ui-view="migrationView"
-        ></div>
-
-        <!-- SERVICE TO DELETE -->
-        <div
-            data-ng-if="!PackMigrationCtrl.loading.init && PackMigrationCtrl.process.currentStep === 'serviceDelete'"
-            data-ui-view="serviceDeleteView"
-        ></div>
-    </div>
-</div>
